+++ src/components/tab-group/tab-group.test.ts
@@ -361,6 +361,31 @@ describe('<wa-tab-group>', () => {
           return expectCustomTabToBeActiveAfter(tabGroup, () => clickOnElement(customHeader!));
         });
 
+        it('selects a tab by changing it via active property', async () => {
+          const tabGroup = await fixture<WaTabGroup>(html`
+            <wa-tab-group>
+              <wa-tab slot="nav" panel="general" data-testid="general-header">General</wa-tab>
+              <wa-tab slot="nav" panel="custom" data-testid="custom-header">Custom</wa-tab>
+              <wa-tab-panel name="general">This is the general tab panel.</wa-tab-panel>
+              <wa-tab-panel name="custom" data-testid="custom-tab-content">This is the custom tab panel.</wa-tab-panel>
+            </wa-tab-group>
+          `);
+
+          const customHeader = queryByTestId<WaTab>(tabGroup, 'custom-header')!;
+          const generalHeader = await waitForHeaderToBeActive(tabGroup, 'general-header');
+          generalHeader.focus();
+
+          expect(customHeader).not.to.have.attribute('active');
+
+          const showEventPromise = oneEvent(tabGroup, 'wa-tab-show') as Promise<WaTabShowEvent>;
+          customHeader.active = true;
+
+          await tabGroup.updateComplete;
+          expect(customHeader).to.have.attribute('active');
+          await expectPromiseToHaveName(showEventPromise, 'custom');
+          return expectOnlyOneTabPanelToBeActive(tabGroup, 'custom-tab-content');
+        });
+
         it('does not change if the active tab is reselected', async () => {
           const tabGroup = await fixture<WaTabGroup>(html`
             <wa-tab-group>